requirements.txt
numpy==1.26.3
pandas==2.2.2
pillow==10.2.0
pytorch-cuda==12.1
scikit-learn==1.5.1
timm==1.0.9
torch==2.6.0
torchvision==0.21.0
tqdm==4.65.0
transformers==4.44.2
